fix BI_randm(), change type of size bit to u16

+++ bigint.h
@@ -24,7 +24,7 @@
  struct bigint_st {
 	u8 value[MAXSIZE + 1];	/* represent the big number */
 	int8_t signbit; 		/* 1 if positive, -1 if negative */
-	u8 size;				/* size in bit */
+	u16 size;				/* size in bit */
  };
 
 /* Inverse a string
@@ -33,7 +33,7 @@
 char* inv_str(const char *str);
 
 //compute size in bits.
-u8 compute_size(BIGINT rop);
+u16 compute_size(BIGINT rop);
 
 /* Initialize a big integer before using 
  * 
@@ -62,7 +62,7 @@ int BI_set_str(BIGINT rop, const char* str, int base);	/*
 														 * The base may vary from values 2 (binary), 10 (decimal) or 16 (hex)
 														 * This function returns 0 if the entire string is a valid number in base base. Otherwise it returns â1.	
 														 */
-void BI_set_pow_2(BIGINT rop, u8 n);	// Set rop = 2^n
+void BI_set_pow_2(BIGINT rop, u16 n);	// Set rop = 2^n
 
 void BI_swap(BIGINT x, BIGINT y);		// swap the values x and y
 
@@ -86,8 +86,8 @@ int64_t BI_get_si(const BIGINT op); /* If op fits into a signed long int return 
 char* BI_get_str(char *str, int base, const BIGINT op); /* Convert op to a string of digits in base base */
 
 /* Random Number Functions */
-void BI_randb(BIGINT rop, u8 n); 			/* Generate a uniformly distributed random integer in the range 0 to 2^nâ1, inclusive. */
-void BI_randm(BIGINT rop, const BIGINT n); 	/* Generate a uniformly distributed random integer in the range 0 to n â 1, inclusive. */
+void BI_randb(BIGINT rop, u16 n); 			/* Generate a uniformly distributed random integer in the range 0 to 2^n-1, inclusive. */
+void BI_randm(BIGINT rop, const BIGINT n); 	/* Generate a uniformly distributed random integer in the range 0 to n - 1, inclusive. */
 
 /* Comparison Functions */
 
@@ -120,9 +120,9 @@ void BI_xor(BIGINT rop, const BIGINT op1, const BIGINT op2); /* Set rop to op1 b
 
 void BI_com(BIGINT rop, const BIGINT op); /* Set rop to the one's complement of op. */
 
-u8 BI_Hammingweight(const BIGINT op); /* Count the number of 1 bits in the binary representation. */
+u16 BI_Hammingweight(const BIGINT op); /* Count the number of 1 bits in the binary representation. */
 
-u8 BI_hamdist(const BIGINT op1, const BIGINT op2); /* Return the hamming distance between the two operands, 
+u16 BI_hamdist(const BIGINT op1, const BIGINT op2); /* Return the hamming distance between the two operands, 
 													* which is the number of bit positions where op1 and op2 have different bit values. 
 													*/
 													
@@ -135,16 +135,16 @@ u8 BI_hamdist(const BIGINT op1, const BIGINT op2); /* Return the hamming distanc
  * If there's no bit found, then the largest possible u8 is returned.
  * 
  */
-u8 BI_scan0(const BIGINT op, u8 starting_bit);
-u8 BI_scan1(const BIGINT op, u8 starting_bit);
+u16 BI_scan0(const BIGINT op, u16 starting_bit);
+u16 BI_scan1(const BIGINT op, u16 starting_bit);
 
-void BI_setbit(BIGINT rop, u8 bit_index); /* Set bit bit_index in rop. */
+void BI_setbit(BIGINT rop, u16 bit_index); /* Set bit bit_index in rop. */
 
-void BI_clrbit(BIGINT rop, u8 bit_index); /* Clear bit bit_index in rop. */
+void BI_clrbit(BIGINT rop, u16 bit_index); /* Clear bit bit_index in rop. */
 
-void BI_combit(BIGINT rop, u8 bit_index); /* Complement bit bit_index in rop. */
+void BI_combit(BIGINT rop, u16 bit_index); /* Complement bit bit_index in rop. */
 
-int BI_tstbit(const BIGINT op, u8 bit_index); /* Test bit bit_index in op and return 0 or 1 accordingly. */
+int BI_tstbit(const BIGINT op, u16 bit_index); /* Test bit bit_index in op and return 0 or 1 accordingly. */
 
 /* Determine whether op is odd or even, respectively. 
  * Return non-zero if yes, zero if no. 
@@ -156,7 +156,7 @@ int BI_even_p(const BIGINT op);
  * The sign of op is ignored, just the absolute value is used. 
  * The result will be either exact or 1 too big. If base is a power of 2, the result is always exact. If op is zero the return value is always 1.
  */
-u8 BI_sizeinbase(const BIGINT op, int base); 
+u16 BI_sizeinbase(const BIGINT op, int base); 
 
 
 /* 
@@ -187,7 +187,7 @@ void BI_submul(BIGINT rop, const BIGINT op1, const BIGINT op2);
 void BI_submul_ui(BIGINT rop, const BIGINT op1, u64 op2);
 
 /* Set rop = op1 x 2 ^ op2. This operation can also be defined as a left shift by op2 bits */
-void BI_shift(BIGINT rop, const BIGINT op1, u8 op2);
+void BI_shift(BIGINT rop, const BIGINT op1, u16 op2);
 
 /* Set rop = op1 / op2. Division is undefined if the divisor is zero */
 void BI_div(BIGINT rop, const BIGINT op1, const BIGINT op2);
